Make stream to output optional #98
Instead of registering a subscriber, the '-v'
command option can be used.
