+++ docker-compose.yml
@@ -108,13 +108,13 @@ services:
     ##############################
     php-connect-four-follow-event-store:
         <<: *php-container
-        command: bin/restartOnChange bin/console connect-four:follow-event-store pointer --select-all-subscribers
+        command: bin/restartOnChange bin/console connect-four:follow-event-store pointer --select-all-subscribers -v
     php-chat-follow-event-store:
         <<: *php-container
-        command: bin/restartOnChange bin/console chat:follow-event-store pointer --select-all-subscribers
+        command: bin/restartOnChange bin/console chat:follow-event-store pointer --select-all-subscribers -v
     php-identity-follow-event-store:
         <<: *php-container
-        command: bin/restartOnChange bin/console identity:follow-event-store pointer --select-all-subscribers
+        command: bin/restartOnChange bin/console identity:follow-event-store pointer --select-all-subscribers -v
     php-consume-messages:
         <<: *php-container
         command: bin/restartOnChange bin/console gaming:consume-messages --select-all-consumers
